可移植性

+++ bv-api/bv-api.h
@@ -4,16 +4,16 @@
 #define _BV_API
 
 #include <stdio.h>
-
+#include <stdint.h>
 void bvInitialize(void);
 // 初始化BV_codec。
 // 可以多次调用。
 
-long long bvDec(char x[]);
+int64_t bvDec(char x[]);
 // 把BV号转换为av号。
 // eg bvDec("BV1xx411c7XW")
 
-char * bvEnc(long long x);
+char * bvEnc(int64_t x);
 // 把av号转换为BV号。
 // eg bvEnc(314LL)
